requirements.txt
amqp==5.3.1
annotated-types==0.6.0
anyio==4.10.0
ariadne==0.19.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.110.0
gcloud==0.18.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-monitoring==2.27.2
google-cloud-pubsub==2.31.1
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-genai==1.31.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphql-core==3.2.6
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
grpcio-tools==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.4
importlib_metadata==8.7.0
kombu==5.5.4
motor==3.3.1
oauth2client==4.1.3
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pymongo==4.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
reportlab==4.4.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.0
starlette==0.36.3
strawberry-graphql==0.276.2
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.28.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
zipp==3.23.0
